requirements.txt
alabaster==0.7.10
app-version==1.0.1
appdirs==1.4.3
APScheduler==3.5.3
asgi-redis==1.4.2
asgiref==1.1.2
asn1crypto==0.24.0
attrs==18.2.0
autobahn==17.8.1
Automat==0.6.0
aylien-apiclient==0.6.0
Babel==2.5.1
backports.functools-lru-cache==1.5
backports.weakref==1.0rc1
bcrypt==3.1.4
beautifulsoup4==4.6.0
bleach==1.5.0
bs4==0.0.1
certifi==2017.7.27.1
cffi==1.11.2
cfscrape==1.8.1
chardet==3.0.4
cmake==0.8.0
colander==1.4
colorama==0.3.9
constantly==15.1.0
construct==2.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.1.4
cssselect==1.0.1
csv2json==0.2
cycler==0.10.0
decorator==4.0.11
defusedxml==0.5.0
dj-core==0.0.4
dj-core-drf==0.0.4
dj-database-url==0.4.2
Django==1.11.6
django-allauth==0.33.0
django-background-tasks==1.1.12
django-braces==1.11.0
django-common-helpers==0.9.1
django-compat==1.0.14
django-cors-headers==2.1.0
django-cron==0.5.0
django-enumfields==0.9.0
django-environ==0.4.4
django-extensions==1.9.0
django-extra-fields==0.9
django-facebook==6.0.3
django-filter==1.0.4
django-jsonfield==1.0.1
django-jwt-auth==0.0.2
django-notify-x==0.1.9
django-oauth-toolkit==1.0.0
django-postgreconnect==0.2
django-ratelimit==1.1.0
django-rest-auth==0.9.1
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.0.8
django-rest-swagger==2.1.2
django-revproxy==0.9.13
django-simple-email-confirmation==0.22
djangorestframework==3.6.4
djangorestframework-filters==0.10.1
djangorestframework-jwt==1.11.0
docutils==0.14
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
enum34==1.1.6
explicit==0.1.3
funcsigs==1.0.2
future==0.15.2
google-api-python-client==1.6.2
gunicorn==19.7.1
h2==3.0.1
h5py==2.7.0
hpack==3.0.0
html5lib==0.9999999
httplib2==0.10.3
hyperframe==5.1.0
hyperlink==17.3.1
idna==2.6
imagesize==0.7.1
incremental==17.5.0
ipaddress==1.0.22
iso8601==0.1.12
itypes==1.1.0
Jinja2==2.9.6
jsonfield==2.0.2
jsonpath-rw==1.4.0
jsonpickle==0.9.5
Keras==2.0.6
lxml==3.8.0
Markdown==2.6.8
MarkupSafe==1.0
mock==2.0.0
msgpack-python==0.4.8
nltk==3.2.2
nose==1.3.7
numpy==1.13.3
oauth2client==4.1.2
oauthlib==2.0.6
openapi-codec==1.3.2
packaging==16.8
pandas==0.20.3
paramiko==2.4.0
parsel==1.3.1
passlib==1.7.1
pbr==3.1.1
pefile==2017.8.1
pluggy==0.5.2
ply==3.10
priority==1.3.0
protobuf==3.3.0
psycopg2==2.7.3.2
py==1.4.34
Py-Authorize==1.4.0.0
pyasn1==0.4.2
pyasn1-modules==0.0.11
pycparser==2.18
PyDispatcher==2.0.5
PyExecJS==1.4.0
Pygments==2.2.0
PyHamcrest==1.9.0
PyJWT==1.5.3
pymongo==3.4.0
PyNaCl==1.2.1
pyOpenSSL==17.2.0
pyparsing==2.2.0
pyquery==1.2.17
python-dateutil==2.6.1
python-firebase==1.2
python-openid==2.2.5
python-ptrace==0.9.2
python-social-auth==0.3.6
python-zimbra==2.0
python3-openid==3.1.0
pytrends==4.2.2
pytz==2017.2
PyXB==1.2.5
PyYAML==3.12
queuelib==1.4.2
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rest-social-auth==1.1.0
retry==0.9.2
rsa==3.4.2
scikit-learn==0.19.1
scipy==1.0.0
selenium==3.8.0
service-identity==17.0.0
simplejson==3.11.1
six==1.11.0
sklearn==0.0
snowballstemmer==1.2.1
social-auth-app-django==0.1.0
social-auth-core==0.2.1
Sphinx==1.6.5
sphinx-rtd-theme==0.2.4
sphinxcontrib-websupport==1.0.1
splinter==0.7.7
Theano==0.9.0
tox==2.9.1
translationstring==1.3
tweepy==3.5.0
twitter==1.17.1
txaio==2.8.1
typing==3.6.2
tzlocal==1.5.1
Unidecode==0.4.21
uritemplate==3.0.0
urllib3==1.22
validate-email==1.3
virtualenv==15.1.0
w3lib==1.19.0
waiter==0.5
webhoseio==0.5
Werkzeug==0.12.2
wget==3.2
whitenoise==3.3.0
XlsxWriter==1.0.2
zope.interface==4.4.2
